Change staging terminology to match common usage
For the record, I still think staging/destaging is better than
staging in and staging out. It's a verbed noun, to stage something
is putting something on a stage. So, we stage the files, then do
something with them on that stage, then destage the results (take
them off the stage again). We're copying, not moving, but that
works.

One normally doesn't place anything inside of a stage, so staging
in isn't right, and staging out is completely weird, and possibly
a contradiction in terms. Looks like at some point people started
using staging as a synonym for copying. Linguistic laziness, bah!

However, if this is what people call it, then this is what we'll
use, because the goal here is to communicate effectively, not to
satisfy pedantic amateur linguists like myself. There.